Create RecoveryManager.sol
RecoveryManager is a partner contract to IdentityManager, allowing further reduction in the cost of deploying a full new identity. 

Note: This currently does not allow a user to change their delegates after they have created a recovery identity within this contract. 

The reasons for this are 1) it dramatically simplifies the code, and 2) allowing for this would require changes to IdentityManager (as adding/removing delegates is the same as changing the recovery, and thus must be subject to the same limitations -- meaning functions must be added to IdentityManager). 

However, a user can still change their delegates, they would just have to create a recovery identity in a different recoveryManager (which would require deploying more than one). Also, should create a "solo" version of this.